removed coursesofferedpersem among other things
