+++ App/models/__init__.py
@@ -3,7 +3,7 @@ from .courses import *
 from .program import *
 from .staff import *
 from .student import *
-from .coursesOfferedPerSem import *
+# from .coursesOfferedPerSem import *
 from .programCourses import *
 from .prerequisites import *
 from .studentCourseHistory import *
