+++ App/models/coursesOfferedPerSem.py
@@ -1,16 +1,16 @@
-from App.database import db 
+# from App.database import db 
 
-class CoursesOfferedPerSem(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    code = db.Column(db.ForeignKey('course.courseCode'))
+# class CoursesOfferedPerSem(db.Model):
+#     id = db.Column(db.Integer, primary_key=True)
+#     code = db.Column(db.ForeignKey('course.courseCode'))
     
-    associated_course = db.relationship('Course', back_populates='offered', overlaps="courses")
+#     # associated_course = db.relationship('Course', back_populates='offered', overlaps="courses")
 
-    def __init__(self, courseCode):
-        self.code = courseCode
+#     def __init__(self, courseCode):
+#         self.code = courseCode
        
-    def get_json(self):
-        return{
-            'ID:': self.id,
-            'Course Code:': self.code
-        }
+#     def get_json(self):
+#         return{
+#             'ID:': self.id,
+#             'Course Code:': self.code
+#         }
