+++ App/models/staff.py
@@ -6,6 +6,8 @@ class Staff(User):
     
     id = db.Column(db.String(10), db.ForeignKey('user.id'), primary_key=True)
     name = db.Column(db.String(50))
+    # courses = db.relationship('StudentCourseHistory', backref='student', lazy=True)
+
 
     def __init__(self, password, staff_id, name):
         super().__init__(staff_id, password)
